+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Features:
 
 - Define the `OpenApiGenerateSpecification` workspace function (`cs openapi genSpec`) and provide its workspace-level implementation, which merges all OpenAPI specifications in a single file.
+- Support the `envFile` options for `DockerService.run`.
 
 ## v0.14.1 (2023-08-04)